refactor: extract GalleryState, optimize image scanning, and standardize model PK
